% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSRforSS-package.R
\docType{package}
\name{CSRforSS-package}
\alias{CSRforSS}
\alias{CSRforSS-package}
\title{CSRforSS: Complex Systems Research for Social Science}
\description{
This packages contains the various R scripts, data, and other files associated with the textbook "Complex Systems Research for Social Science."
}
\author{
\strong{Maintainer}: Bernard Ricca \email{barneyricca@gmail.com} (\href{https://orcid.org/0000-0001-6318-1222}{ORCID})

}
\keyword{internal}
