% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chapter_Scripts.R
\name{test_add}
\alias{test_add}
\title{test_add}
\usage{
test_add(a, b)
}
\arguments{
\item{<a>}{first addend}

\item{<b>}{second addend}
}
\description{
This function returns the sum of two numbers.
}
\examples{
test_add(2,3)
}
\keyword{add}
